+++ smile_id_core/IdApi.py
@@ -1,8 +1,6 @@
 """ID API class for kyc services."""
 from typing import Any, Dict, Optional, Union
 
-from requests import Response
-
 from smile_id_core.base import Base
 from smile_id_core.BusinessVerification import BusinessVerification
 from smile_id_core.constants import JobType
@@ -39,7 +37,7 @@ class IdApi(Base):
         partner_params: Dict[str, Any],
         id_params: Dict[str, str],
         options_params: Optional[OptionsParams] = None,
-    ) -> Response:
+    ) -> Dict[str, Any]:
         """Validate data params & query id_verification endpoint for KYC jobs.
 
         Performs checks on id_info_params, partner_params, makes endpoint
@@ -94,4 +92,4 @@ class IdApi(Base):
                 f"Failed to post entity to {url},"
                 f" status={response.status_code}, response={response.json()}"
             )
-        return response
+        return dict(response.json())