+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Changed
-
+- Breaking: Changed return types of methods from `Response` to `Dict[str,Any]`
 - Breaking: Change `submit_job` method signature for both `WebApi` and `IDApi`.
   `WebApi` and `IDApi` will no longer support `use_validation_api` parameter. Migration guide available [here](migration_guides/v2_2_0.md)